DMP'25/Tornado-Server-Invoice-App/requirements.txt
tornado>=6.0.0
PyMySQL>=1.0.0
dropbox>=11.36.0
python-memcached>=1.59
pytest>=7.0.0
pytest-tornado>=0.8.1
cryptography>=3.4.8
passlib>=1.7.4
boto3>=1.20.0
python-dotenv>=0.19.0
PyJWT>=2.0.0
pdfkit>=1.0.0
beautifulsoup4>=4.9.0